Extract Mapbox logic into reusable hook
